+++ NonlocalGameValue.m
@@ -1,4 +1,4 @@
-%%  NONLOCALGAMEVALUE    Computes the maximum value of a two-player nonlocal game
+%%  NONLOCALGAMEVALUE    Computes the maximum value of a two-player non-local game
 %   This function has two required input arguments:
 %     P: a matrix whose (x,y)-entry is the probability that the referee
 %        asks Alice question x and Bob question y
@@ -7,25 +7,28 @@
 %        and y.
 %
 %   NGVAL = NonlocalGameValue(P,V) is the maximum value that the specified 
-%   nonlocal game can take on in classical mechanics. For the maximum
+%   non-local game can take on in classical mechanics. For the maximum
 %   quantum or no-signalling value, see the optional arguments described
 %   below.
 %
-%   This function has two optional input arguments:
+%   This function has three optional input arguments:
 %     MTYPE (default 'classical'): one of 'classical', 'quantum', or
 %       'nosignal', indicating what type of nonlocal game value should be
 %       computed. IMPORTANT NOTE: if MTYPE='quantum' then only an upper
-%       bound on the nonlocal game value is computed, not its exact value
+%       bound on the non-local game value is computed, not its exact value
 %       (see the argument K below).
 %     K (default 1): if MYTPE='quantum', then K is a non-negative integer
 %       or string indicating what level of the NPA hierarchy to use to
-%       bound the nonlocal game (higher values give better bounds, but
+%       bound the non-local game (higher values give better bounds, but
 %       require more computation time). See the NPAHierarchy function for
-%       details.
+%       details. If MTYPE is anything else, then K is simply ignored.
+%     REPT (default 1): the number of times that the non-local game should
+%       be played in parallel (i.e., the number of repetitions of the
+%       game).
 %
-%   NGVAL = NonlocalGameValue(P,V,MTYPE,K) is the maximum value that the
-%   specified nonlocal game can take on in the setting (classical, quantum,
-%   or no-signalling) specified by MTYPE.
+%   NGVAL = NonlocalGameValue(P,V,MTYPE,K,REPT) is the maximum value that
+%   the specified non-local game can take on in the setting (classical,
+%   quantum or no-signalling) specified by MTYPE.
 %
 %   URL: http://www.qetlab.com/NonlocalGameValue
 
@@ -34,17 +37,49 @@
 %
 %   author: Nathaniel Johnston (nathaniel@njohnston.ca)
 %   package: QETLAB
-%   last updated: March 6, 2015
+%   last updated: July 10, 2015
 
 function ngval = NonlocalGameValue(p,V,varargin)
 
-    % set optional argument defaults: MTYPE='classical', K=1
-    [mtype,k] = opt_args({ 'classical', 1 },varargin{:});
+    % set optional argument defaults: MTYPE='classical', K=1, REPT=1
+    [mtype,k,rept] = opt_args({ 'classical', 1, 1 },varargin{:});
 
     % Get some basic values.
     [ma,mb] = size(p);
     oa = size(V,1);
     ob = size(V,2);
+    
+    % Are we playing the game multiple times? If so, adjust p and V
+    % accordingly.
+    if(rept > 1)
+        % Create the new probability array.
+        p = Tensor(p,rept);
+        
+        % Create the new winning output array (this is more complicated
+        % since MATLAB doesn't have any built-in functions for tensoring
+        % together 4D arrays).
+        V2 = zeros(oa^rept,ob^rept,ma^rept,mb^rept);
+        i_ind = zeros(1,rept);
+        j_ind = zeros(1,rept);
+        for i = 1:ma^rept
+            for j = 1:mb^rept
+                for l = rept:-1:1
+                    to_tensor{l} = V(:,:,i_ind(l)+1,j_ind(l)+1);
+                end
+                V2(:,:,i,j) = Tensor(to_tensor);
+
+                j_ind = update_odometer(j_ind,mb*ones(1,rept));
+            end
+            i_ind = update_odometer(i_ind,ma*ones(1,rept));
+        end
+        V = V2;
+
+        % Update some computed values.
+        ma = ma^rept;
+        mb = mb^rept;
+        oa = oa^rept;
+        ob = ob^rept;
+    end
     
     % The no-signalling maximum is just implemented by taking the zero-th
     % level of the NPA (quantum) hierarchy.
@@ -79,8 +114,8 @@ function ngval = NonlocalGameValue(p,V,varargin)
         ngval = -Inf;
         a_ind = zeros(1,ma);
         b_ind = zeros(1,mb);
-        a_ind(ma)=oa-2;
-        b_ind(mb)=ob-2;
+        a_ind(ma) = oa-2;
+        b_ind(mb) = ob-2;
         p = permute(repmat(p,[1,1,oa,ob]),[3,4,1,2]);
         ab_prob = zeros(oa,ob,ma,mb);